Todo: 7
Continuing work on the API. Added default user list logic, CRUD methods for the lists and their individual tasks. Specified logic for a default user list to be created, if it's not already present in the database for the user. This allows for every new task that is created without an appointed list will be saved on that list by default.

Co-Authored-By: arivepr <xanverapr@gmail.com>